+++ lib/rann/product_neuron.rb
@@ -5,7 +5,7 @@ module RANN
     attr_accessor :intermediate
 
     def set_value!
-      @intermediate = incoming.reduce{ |i, m| m * i }
+      @intermediate = incoming.reduce :*
       self.value    = ACTIVATION_FUNCTIONS[activation_function].call @intermediate
     end
   end